Creating WorkerPool and moving last functionality out of supervisor
