+++ mephisto/operations/client_io_handler.py
@@ -29,7 +29,7 @@ from mephisto.abstractions.blueprint import AgentState
 from mephisto.data_model.agent import Agent, OnboardingAgent
 from mephisto.operations.datatypes import LiveTaskRun, ChannelInfo, AgentInfo
 from mephisto.abstractions.channel import Channel, STATUS_CHECK_TIME
-from typing import Dict, Set, Optional, List, Any, Union, TYPE_CHECKING
+from typing import Dict, Set, Tuple, Optional, List, Any, Union, TYPE_CHECKING
 
 if TYPE_CHECKING:
     from mephisto.abstractions.database import MephistoDB
@@ -180,6 +180,8 @@ class ClientIOHandler:
 
     def _on_act(self, packet: Packet, _channel_id: str):
         """Handle an action as sent from an agent, enqueuing to the agent"""
+        live_run = self._live_run
+        assert live_run is not None, "Must have initialized a live run first"
         agent = self._get_agent_for_id(packet.sender_id)
 
         # If the packet is_submit, and has files, we need to
@@ -188,7 +190,7 @@ class ClientIOHandler:
             data_files = packet.data.get("files")
             if data_files is not None:
                 save_dir = agent.get_data_dir()
-                architect = self._live_run.architect
+                architect = live_run.architect
                 for f_obj in data_files:
                     architect.download_file(f_obj["filename"], save_dir)
 
@@ -197,6 +199,8 @@ class ClientIOHandler:
 
     def _register_agent(self, packet: Packet, channel_id: str) -> None:
         """process the packet, then delegate to the worker pool appropriate registration"""
+        live_run = self._live_run
+        assert live_run is not None, "Must have initialized a live run first"
         request_id = packet.data["request_id"]
         self.request_id_to_channel_id[request_id] = channel_id
         crowd_data = packet.data["provider_data"]
@@ -212,14 +216,16 @@ class ClientIOHandler:
             )
             return
 
-        self._REPLACE_SUPERVISOR.register_agent(crowd_data, request_id, self._live_run)
+        live_run.worker_pool.register_agent(crowd_data, request_id, self._live_run)
 
     def _register_worker(self, packet: Packet, channel_id: str) -> None:
         """Read and forward a worker registration packet"""
+        live_run = self._live_run
+        assert live_run is not None, "Must have initialized a live run first"
         request_id = packet.data["request_id"]
         self.request_id_to_channel_id[request_id] = channel_id
         crowd_data = packet.data["provider_data"]
-        self._REPLACE_SUPERVISOR.register_worker(crowd_data, request_id, self._live_run)
+        live_run.worker_pool.register_worker(crowd_data, request_id)
 
     def send_provider_details(self, request_id: str, additional_data: Dict[str, Any]):
         base_data = {"request_id": request_id}
@@ -237,7 +243,9 @@ class ClientIOHandler:
 
     def _get_init_data(self, packet, channel_id: str):
         """Get the initialization data for the assigned agent's task"""
-        task_runner = self._live_run.task_runner
+        live_run = self._live_run
+        assert live_run is not None, "Must have initialized a live run first"
+        task_runner = live_run.task_runner
         agent_id = packet.data["provider_data"]["agent_id"]
         agent = self._get_agent_for_id(agent_id)
         assert isinstance(
@@ -263,8 +271,10 @@ class ClientIOHandler:
 
     def _on_submit_onboarding(self, packet: Packet, channel_id: str) -> None:
         """Handle the submission of onboarding data"""
+        live_run = self._live_run
+        assert live_run is not None, "Must have initialized a live run first"
         onboarding_id = packet.sender_id
-        if onboarding_id not in self._REPLACE_SUPERVISOR.agents:
+        if onboarding_id not in live_run.worker_pool.agents:
             logger.warning(
                 f"Onboarding agent {onboarding_id} already submitted or disconnected, "
                 f"but is calling _on_submit_onboarding again"
@@ -287,6 +297,8 @@ class ClientIOHandler:
 
     def _on_message(self, packet: Packet, channel_id: str):
         """Handle incoming messages from the channel"""
+        live_run = self._live_run
+        assert live_run is not None, "Must have initialized a live run first"
         # TODO(#102) this method currently assumes that the packet's sender_id will
         # always be a valid agent in our list of agent_infos. At the moment this
         # is a valid assumption, but will not be on recovery from catastrophic failure.
@@ -302,7 +314,7 @@ class ClientIOHandler:
             self._get_init_data(packet, channel_id)
         elif packet.type == PACKET_TYPE_RETURN_AGENT_STATUS:
             # Record this status response
-            self._REPLACE_SUPERVISOR._handle_updated_agent_status(packet.data)
+            live_run.worker_pool.handle_updated_agent_status(packet.data)
         elif packet.type == PACKET_TYPE_ERROR_LOG:
             self._log_frontend_error(packet)
         else:
@@ -427,7 +439,9 @@ class ClientIOHandler:
 
     def _get_agent_for_id(self, agent_id: str):
         """Temporary method to get an agent, while API is figured out"""
-        return self._REPLACE_SUPERVISOR.agents[agent_id].agent
+        live_run = self._live_run
+        assert live_run is not None, "Must have initialized a live run first"
+        return live_run.worker_pool.agents[agent_id].agent
 
     def _get_channel_for_agent(self, agent_id: str) -> Channel:
         """Return the sending channel for a given agent"""
